make a lib out of the trigger
